added CoursesCardListComponent
