+++ udemy-1-reactive-angular-course/src/app/app.module.ts
@@ -33,8 +33,10 @@ import {SafeUrlPipe} from './common/safe-url.pipe';
 import {MessagesComponent} from './messages/messages.component';
 import {SearchLessonsComponent} from './search-lessons/search-lessons.component';
 import { LoadingComponent } from './loading/loading.component';
+import { CoursesCardListComponent } from './courses-card-list/courses-card-list.component';
 
-@NgModule({ declarations: [
+@NgModule({
+    declarations: [
         AppComponent,
         HomeComponent,
         AboutComponent,
@@ -45,9 +47,11 @@ import { LoadingComponent } from './loading/loading.component';
         SafeUrlPipe,
         MessagesComponent,
         SearchLessonsComponent,
-        LoadingComponent
+        LoadingComponent,
+        CoursesCardListComponent
     ],
-    bootstrap: [AppComponent], imports: [BrowserModule,
+    bootstrap: [AppComponent],
+    imports: [BrowserModule,
         BrowserAnimationsModule,
         MatMenuModule,
         MatButtonModule,
@@ -67,7 +71,8 @@ import { LoadingComponent } from './loading/loading.component';
         MatSelectModule,
         MatDatepickerModule,
         MatMomentDateModule,
-        ReactiveFormsModule], providers: [
+        ReactiveFormsModule],
+    providers: [
         provideHttpClient(withInterceptorsFromDi())
     ] })
 export class AppModule {
