+++ udemy-1-reactive-angular-course/src/app/courses-card-list/courses-card-list.component.html
@@ -1,0 +1,24 @@
+<mat-card *ngFor="let course of courses" class="course-card mat-elevation-z10">
+
+    <mat-card-header>
+        <mat-card-title>{{course.description}}</mat-card-title>
+    </mat-card-header>
+
+    <img mat-card-image [src]="course.iconUrl">
+
+    <mat-card-content>
+        <p>{{course.longDescription}}</p>
+    </mat-card-content>
+
+    <mat-card-actions class="course-actions">
+
+        <button mat-button mat-raised-button color="primary" [routerLink]="['/courses', course.id]">
+            VIEW COURSE
+        </button>
+
+        <button mat-button mat-raised-button color="accent" (click)="editCourse(course)">
+            EDIT
+        </button>
+
+    </mat-card-actions>
+</mat-card>
