+++ udemy-1-reactive-angular-course/src/app/courses-card-list/courses-card-list.component.scss
@@ -1,5 +1,3 @@
-
-
 .course-card {
     margin: 20px 10px;
 }
@@ -12,4 +10,3 @@
 .course-actions button {
   margin-right: 5px;
 }
-
