+++ udemy-1-reactive-angular-course/src/app/courses-card-list/courses-card-list.component.ts
@@ -1,0 +1,31 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { Course } from '../model/course';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { CourseDialogComponent } from '../course-dialog/course-dialog.component';
+
+@Component({
+  selector: 'courses-card-list',
+  standalone: false,
+  templateUrl: './courses-card-list.component.html',
+  styleUrl: './courses-card-list.component.scss'
+})
+export class CoursesCardListComponent {
+  constructor(private dialog: MatDialog) {
+  }
+
+  @Input()
+  courses: Course[];
+
+  editCourse(course: Course) {
+
+    const dialogConfig = new MatDialogConfig();
+
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.width = "400px";
+
+    dialogConfig.data = course;
+
+    const dialogRef = this.dialog.open(CourseDialogComponent, dialogConfig);
+  }
+}
