+++ udemy-1-reactive-angular-course/src/app/home/home.component.css
@@ -1,26 +1,8 @@
-
-
 .courses-panel {
     max-width: 400px;
     margin: 0 auto;
 }
 
-
-
-
 .course-card {
   margin: 20px 10px;
 }
-
-.course-actions {
-  display: flex;
-  justify-content: center;
-}
-
-
-.course-actions button {
-  display: flex;
-  justify-content: center;
-  margin-right: 5px;
-}
-
