+++ udemy-1-reactive-angular-course/src/app/home/home.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Course, sortCoursesBySeqNo } from '../model/course';
 import { map } from 'rxjs/operators';
-import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { CourseDialogComponent } from '../course-dialog/course-dialog.component';
 import { CoursesService } from '../services/courses.service';
 import { Observable } from 'rxjs';
 
@@ -17,8 +15,7 @@ export class HomeComponent implements OnInit {
   advancedCourses$: Observable<Course[]>;
 
   constructor(
-    private coursesService: CoursesService,
-    private dialog: MatDialog) {
+    private coursesService: CoursesService) {
   }
 
   ngOnInit() {
@@ -30,18 +27,7 @@ export class HomeComponent implements OnInit {
       map(courses => courses.filter(course => course.category === "BEGINNER")));
     this.advancedCourses$ = allCourses$.pipe(
       map(courses => courses.filter(course => course.category === "ADVANCED")));
-  }
 
-  editCourse(course: Course) {
-
-    const dialogConfig = new MatDialogConfig();
-
-    dialogConfig.disableClose = true;
-    dialogConfig.autoFocus = true;
-    dialogConfig.width = "400px";
-
-    dialogConfig.data = course;
-
-    const dialogRef = this.dialog.open(CourseDialogComponent, dialogConfig);
+    console.log('HomeComponent initialized!')
   }
 }
